Scope /posts redirects to posts, move tests to own controller test
