+++ test/controllers/blogs/posts_controller_test.rb
@@ -134,112 +134,6 @@ class Blogs::PostsControllerTest < ActionDispatch::IntegrationTest
     assert_equal post, assigns(:post)
   end
 
-  test "should redirect prefixed posts path to canonical post URL on custom domain" do
-    @blog = blogs(:annie)
-    host! @blog.custom_domain
-    post = @blog.posts.visible.first
-
-    get "/posts/#{post.slug}"
-
-    assert_redirected_to "http://#{@blog.custom_domain}/#{post.slug}"
-    assert_equal 301, @response.status
-  end
-
-  test "should redirect prefixed posts path for pages on custom domain" do
-    @blog = blogs(:annie)
-    host! @blog.custom_domain
-    page = posts(:annie_home_page)
-
-    get "/posts/#{page.slug}"
-
-    assert_redirected_to "http://#{@blog.custom_domain}/#{page.slug}"
-    assert_equal 301, @response.status
-  end
-
-  test "should redirect prefixed posts path on pagecord subdomain" do
-    post = @blog.posts.visible.first
-
-    get "/posts/#{post.slug}"
-
-    assert_redirected_to "http://#{@blog.subdomain}.example.com/#{post.slug}"
-    assert_equal 301, @response.status
-  end
-
-  test "should let custom domain redirect take precedence for prefixed posts path" do
-    @blog = blogs(:annie)
-    host_subdomain! @blog.subdomain
-    post = @blog.posts.visible.first
-
-    get "/posts/#{post.slug}"
-
-    assert_redirected_to "http://#{@blog.custom_domain}/posts/#{post.slug}"
-    assert_equal 301, @response.status
-  end
-
-  test "should not redirect prefixed posts path for draft posts" do
-    post = @blog.posts.create!(title: "Draft Post", content: "Draft content", status: :draft)
-
-    get "/posts/#{post.slug}"
-
-    assert_response :not_found
-  end
-
-  test "should redirect prefixed posts path for hidden posts" do
-    post = @blog.posts.create!(
-      title: "Hidden Post",
-      content: "This is hidden content",
-      status: :published,
-      hidden: true
-    )
-
-    get "/posts/#{post.slug}"
-
-    assert_redirected_to "http://#{@blog.subdomain}.example.com/#{post.slug}"
-    assert_equal 301, @response.status
-  end
-
-  test "should not redirect prefixed posts path for discarded posts" do
-    post = @blog.posts.visible.first
-    post.discard!
-
-    get "/posts/#{post.slug}"
-
-    assert_response :not_found
-  end
-
-  test "should not redirect prefixed posts path for future posts" do
-    post = @blog.posts.create!(
-      title: "Future Post",
-      content: "Future content",
-      status: :published,
-      published_at: 1.day.from_now
-    )
-
-    get "/posts/#{post.slug}"
-
-    assert_response :not_found
-  end
-
-  test "should not redirect prefixed posts path for missing content" do
-    get "/posts/missing"
-
-    assert_response :not_found
-  end
-
-  test "should keep posts archive route before prefixed posts redirect" do
-    get "/posts"
-
-    assert_response :success
-    assert_template "blogs/posts/index"
-  end
-
-  test "should keep embedded posts route before prefixed posts redirect" do
-    get "/posts/embedded", params: { style: "card", frame_id: "posts" }
-
-    assert_response :success
-    assert_template "blogs/embedded_posts/index"
-  end
-
   test "show should render full content without excerpt marker" do
     post = @blog.posts.create!(
       title: "Excerpted Show Post",
@@ -1423,10 +1317,6 @@ class Blogs::PostsControllerTest < ActionDispatch::IntegrationTest
   end
 
   private
-
-    def host_subdomain!(name)
-      host! "#{name}.#{Rails.application.config.x.domain}"
-    end
 
     def create_content_with_attachment(blog:, title:, caption:, is_page: false)
       blob = ActiveStorage::Blob.create_and_upload!(
