+++ test/support/host_helpers.rb
@@ -1,0 +1,5 @@
+class ActionDispatch::IntegrationTest
+  def host_subdomain!(name)
+    host! "#{name}.#{Rails.application.config.x.domain}"
+  end
+end
